Custom validations

+++ src/main/java/com/betasquirrel/controller/UsersController.java
@@ -1,19 +1,14 @@
 package com.betasquirrel.controller;
 
-import com.betasquirrel.model.ObjectResponse;
 import com.betasquirrel.model.ListResponse;
+import com.betasquirrel.model.ObjectResponse;
 import com.betasquirrel.model.User;
 import com.betasquirrel.repository.UserRepository;
-import com.betasquirrel.util.ValidationError;
-import com.betasquirrel.util.ValidationErrorBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
-import javax.validation.Valid;
 import java.util.List;
 
 /**
@@ -50,7 +45,7 @@ public class UsersController {
      * @return
      */
     @RequestMapping(method = RequestMethod.POST)
-    public ListResponse saveUser(@Valid @RequestBody final User user, HttpServletResponse http) {
+    public ListResponse saveUser(@RequestBody final User user, HttpServletResponse http) {
         userRepository.save(user);
         ListResponse response = new ListResponse();
         response.setMessage("Successfully Created");
@@ -69,10 +64,15 @@ public class UsersController {
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     public ObjectResponse getUser(@PathVariable("id") Integer userId, HttpServletResponse http) {
         ObjectResponse response = new ObjectResponse();
-        response.setMessage("Successfully Retrieved");
-        response.setStatusCode(http.getStatus());
-        User user = userRepository.findOne(userId);
-        response.setData(user);
+        if (userRepository.exists(userId)) {
+            response.setMessage("Successfully Retrieved");
+            response.setStatusCode(http.getStatus());
+            response.setData(userRepository.findOne(userId));
+        } else {
+            response.setMessage("Record not found");
+            response.setStatusCode(404);
+            response.setData(null);
+        }
         return response;
     }
 
@@ -84,15 +84,18 @@ public class UsersController {
      */
     @RequestMapping(method = RequestMethod.PUT)
     public ObjectResponse updateUser(@RequestBody final User user, HttpServletResponse http) {
+        ObjectResponse response = new ObjectResponse();
         if (userRepository.exists(user.getId())) {
             userRepository.updateUser(user.getName(), user.getEmail(), user.getMobile(), user.getId());
-            ObjectResponse response = new ObjectResponse();
             response.setMessage("Successfully Updated");
             response.setStatusCode(http.getStatus());
             response.setData(userRepository.findOne(user.getId()));
-            return response;
+        } else {
+            response.setMessage("Record not found");
+            response.setStatusCode(404);
+            response.setData(null);
         }
-        return null;
+        return response;
     }
 
     /**
@@ -103,22 +106,17 @@ public class UsersController {
      */
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ListResponse deleteUser(@PathVariable("id") Integer userId, HttpServletResponse http) {
-        userRepository.delete(userId);
         ListResponse response = new ListResponse();
-        response.setMessage("Successfully Deleted");
-        response.setStatusCode(http.getStatus());
+        if(userRepository.exists(userId)) {
+            userRepository.delete(userId);
+            response.setStatusCode(http.getStatus());
+            response.setMessage("Successfully Deleted");
+        } else {
+            response.setStatusCode(404);
+            response.setMessage("Record not found");
+        }
         List<User> users = userRepository.findAll();
         response.setData(users);
         return response;
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-    public ValidationError handleException(MethodArgumentNotValidException exception) {
-        return createValidationError(exception);
-    }
-
-    private ValidationError createValidationError(MethodArgumentNotValidException e) {
-        return ValidationErrorBuilder.fromBindingErrors(e.getBindingResult());
     }
 }
